change input type to int

+++ 0x01-python_async_function/0-basic_async_syntax.py
@@ -5,10 +5,8 @@ import asyncio
 import random
 
 
-async def wait_random(max_delay: float = 10) -> float:
+async def wait_random(max_delay: int = 10) -> float:
     """wait for a random number of sec and return the number of sec waited"""
     i: float = random.uniform(0, max_delay)
-
     await asyncio.sleep(i)
-
     return i
